requirements.txt
APScheduler>=3.10
dbutils>=3.0
PyMySQL>=1.1
cryptography>=41
numpy>=1.26
pandas>=2.2
plotly>=5.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
html5lib>=1.1
Jinja2>=3.1
PySocks>=1.7
idna>=3.6
certifi>=2024.2.2
charset-normalizer>=3.3
python-dotenv